Sort bootstrap assets by SHA-256 hash
I'm not sure if an `8-bit/rest` arrangement is right considering the scale, maybe an `8-bit/8-bit/rest` arrangement is better, but we'll make do with this for now.
Besides, we'll probably want a script for the purpose of fetching from an existing cloud or a user cache at least for the time being, as this isn't enough assets to get the engine booted up from a fresh cache, it's only enough for a *warm* cache.

...seriously though, why do we have to keep redownloading this crap? This engine has definitely been a great example of why you should NOT shove everything in the cloud.